@hazzik 2019-01-14
I think the impact is not so significant for this type of a change.