@bahusoid 2019-01-15
> Fetch took: 93.425.240,00ms

Does it mean that fetch took almost 26  hours? :confused: